requirements.txt
crewai
apify-client
python-dotenv
requests
gTTS
mutagen
